Adds start of stripe implementation, Moved a lot of files to their own directories
