+++ app/Domits/src/features/bookingengine/overview/utils/CalculateNumberOfNights.js
